Add support for COLUMNS to adjust calculated terminal width
Some terminals are just not playing nice with the calculations tabular
output does to fit in the terminal.

With this change we add support for a COLUMNS environment variable
that can be set to force the calculations to assume the given number
of columns.

Set with: export COLUMNS=80

or some other number to fix issues displaying tables